backend/requirements.txt
Django==3.2.3
djangorestframework==3.15.1
djoser==2.2.3
drf-extra-fields==3.7.0
Pillow==9.3.0
psycopg2-binary==2.9.3
django-filter==23.1